US-296: Creado servicio de carga de Agendas

+++ grails-app/domain/ordertracker/AgendaDia.groovy
@@ -2,7 +2,7 @@ package ordertracker
 
 class AgendaDia {
     int codigoDia
-    ArrayList<Integer> listaClientes = []
+    ArrayList<Long> listaClientes = []
 
     static belongsTo = [agenda: Agenda]
